requirements/test.txt
pytest
factory_boy
pytest-django
